requirements.txt
# Core dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # Excel file handling

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0  # Better PDF text extraction
pypdf>=3.17.0

# Report generation
jinja2>=3.1.0
weasyprint>=60.0  # HTML to PDF
markdown>=3.5.0

# Web interface (optional - for Streamlit version)
streamlit>=1.30.0
plotly>=5.18.0  # Interactive charts

# File handling
python-magic>=0.4.27  # File type detection
pillow>=10.0.0  # Image processing

# API & utilities
requests>=2.31.0
pydantic>=2.5.0  # Data validation
stripe>=14.0.0  # Payment processing

